+++ app/src/main/java/com/karl/android/coincounter/NewSaveActivity.java
@@ -1,0 +1,19 @@
+package com.karl.android.coincounter;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+
+public class NewSaveActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        Toolbar toolbar = (Toolbar)findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        setContentView(R.layout.activity_new_save);
+    }
+}
+