+++ app/src/main/java/com/karl/android/fragments/MySavesFragment.java
@@ -1,15 +1,25 @@
 package com.karl.android.fragments;
 
 import android.app.Fragment;
+import android.database.Cursor;
 import android.os.Bundle;
+import android.support.v7.widget.DefaultItemAnimator;
+import android.support.v7.widget.DividerItemDecoration;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ListView;
 import android.widget.TextView;
 
 import com.karl.android.coincounter.R;
+import com.karl.android.coincounter.adapters.lists.SavesAdapter;
 import com.karl.android.templates.SlimSave;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import data.sql.MySQLiteHelper;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -20,8 +30,11 @@ public class MySavesFragment extends Fragment {
 
     View rootView;
 
-    ListView savesList;
+    RecyclerView savesList;
     TextView noInformationToDisplay;
+    List<SlimSave> saves = new ArrayList<>();
+
+    SavesAdapter adapter;
 
     public MySavesFragment() {}
 
@@ -33,6 +46,10 @@ public class MySavesFragment extends Fragment {
         // initialise the views of the fragment.
         initialiseViews();
 
+        setUpRecyclerView();
+
+        getInformationFromDatabase();
+
         savesList.setVisibility(View.GONE);
 
         return rootView;
@@ -43,10 +60,22 @@ public class MySavesFragment extends Fragment {
      */
     private void initialiseViews(){
         // lists
-        savesList = (ListView) rootView.findViewById(R.id.my_saves_fragment_list_view);
+        savesList = (RecyclerView) rootView.findViewById(R.id.my_saves_fragment_list_view);
 
         // text views
         noInformationToDisplay = (TextView) rootView.findViewById(R.id.my_saves_fragment_no_information_text_view);
+    }
+
+    /**
+     * Set up the default settings of the recycler view.
+     */
+    private void setUpRecyclerView(){
+        adapter = new SavesAdapter(saves);
+        RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getActivity());
+        savesList.setLayoutManager(layoutManager);
+        savesList.addItemDecoration(new DividerItemDecoration(getActivity(), LinearLayoutManager.VERTICAL));
+        savesList.setItemAnimator(new DefaultItemAnimator());
+        savesList.setAdapter(adapter);
     }
 
     /**
@@ -59,7 +88,26 @@ public class MySavesFragment extends Fragment {
     private SlimSave getInformationFromDatabase(){
         SlimSave save = new SlimSave();
 
-        // get information from the database and add it to the save object.
+        MySQLiteHelper mySQLiteHelper = new MySQLiteHelper(getActivity());
+
+        Cursor totalsFromDatabase = mySQLiteHelper.getDataFromTotals();
+
+        if(totalsFromDatabase.getCount() == 0){
+            noInformationToDisplay.setVisibility(View.VISIBLE);
+            savesList.setVisibility(View.GONE);
+        } else {
+            noInformationToDisplay.setVisibility(View.GONE);
+            savesList.setVisibility(View.VISIBLE);
+            while(totalsFromDatabase.moveToNext()){
+                SlimSave slimSave = new SlimSave();
+                slimSave.setId(totalsFromDatabase.getString(0));
+                slimSave.setTitle(totalsFromDatabase.getString(1));
+                slimSave.setAmount(totalsFromDatabase.getString(2));
+                slimSave.setDate(totalsFromDatabase.getString(3));
+            }
+        }
+
+        adapter.notifyDataSetChanged();
 
         return save;
     }